Moved fixture to a separate file
