+++ tests/end_to_end/test_suites/sample_tests.py
@@ -4,6 +4,7 @@
 import pytest
 import logging
 
+from tests.end_to_end.utils.common_fixtures import fx_federation
 from tests.end_to_end.utils import federation_helper as fed_helper
 
 log = logging.getLogger(__name__)
